Fix compatibility with Django 2.1 with widget.render().
As discussed in https://docs.djangoproject.com/en/2.1/releases/2.1/,
support for a widget.render() function without a renderer argument was
removed.  The behavior if you pass `None` as the renderer is the same
as the old behaviour, so nothing complex is required for this migration.

Commit message rewritten by Tim Abbott.